added extended fields to d3 basic tree structure
list of opening groups, how many times a particular node is visited in
creation of the tree (i.e. node frequency), what opening groups the
move is part of, the board state as a FEN string if the node is at the
end of an opening the name of that opening.